refactor & combine ucd-diskio & ucd-mib
