+++ lang/en/settings.php
@@ -765,7 +765,7 @@ return [
             'stp' => [
                 'description' => 'STP',
             ],
-            'ucd-diskio' => [
+            'ucd-mib' => [
                 'description' => 'UCD DiskIO',
             ],
             'vlans' => [
@@ -1647,9 +1647,6 @@ return [
             ],
             'junose-atm-vp' => [
                 'description' => 'JunOS ATM VP',
-            ],
-            'ucd-diskio' => [
-                'description' => 'UCD DiskIO',
             ],
             'wireless' => [
                 'description' => 'Wireless',
