+++ lang/fr/settings.php
@@ -469,7 +469,7 @@ return [
             'stp' => [
                 'description' => 'STP',
             ],
-            'ucd-diskio' => [
+            'ucd-mib' => [
                 'description' => 'UCD DiskIO',
             ],
             'vlans' => [
@@ -808,9 +808,6 @@ return [
             ],
             'junose-atm-vp' => [
                 'description' => 'JunOS ATM VP',
-            ],
-            'ucd-diskio' => [
-                'description' => 'UCD DiskIO',
             ],
             'wireless' => [
                 'description' => 'Wireless',
