+++ lang/ru/settings.php
@@ -702,7 +702,7 @@ return [
             'stp' => [
                 'description' => 'STP',
             ],
-            'ucd-diskio' => [
+            'ucd-mib' => [
                 'description' => 'UCD DiskIO',
             ],
             'vlans' => [
@@ -1307,9 +1307,6 @@ return [
             ],
             'junose-atm-vp' => [
                 'description' => 'JunOS ATM VP',
-            ],
-            'ucd-diskio' => [
-                'description' => 'UCD DiskIO',
             ],
             'wireless' => [
                 'description' => 'Беспроводная связь',
