+++ lang/uk/settings.php
@@ -563,7 +563,7 @@ return [
             'toner' => [
                 'description' => 'Toner',
             ],
-            'ucd-diskio' => [
+            'ucd-mib' => [
                 'description' => 'UCD DiskIO',
             ],
             'vlans' => [
@@ -1063,9 +1063,6 @@ return [
             ],
             'toner' => [
                 'description' => 'Toner',
-            ],
-            'ucd-diskio' => [
-                'description' => 'UCD DiskIO',
             ],
             'wireless' => [
                 'description' => 'Бездротові',
