+++ resources/definitions/os_detection/adva-fsp150cp.yaml
@@ -16,4 +16,4 @@ discovery_modules:
     storage: false
     hr-device: false
     arp-table: false
-    ucd-diskio: false
+    ucd-mib: false
