+++ resources/definitions/os_detection/aviat-wtm.yaml
@@ -19,7 +19,7 @@ discovery:
             - .1.3.6.1.4.1.2509.11.1.1.15
 discovery_modules:
     ucd-dsktable: false
-    ucd-diskio: false
+    ucd-mib: false
     arp-table: false
     fdb-table: false
     vlans: false
