+++ resources/definitions/os_detection/coriant.yaml
@@ -6,7 +6,6 @@ discovery_modules:
     ntp: false
     ospf: false
     stp: false
-    ucd-diskio: false
     ucd-mib: false
     ipmi: false
     bgp-peers: false
@@ -20,7 +19,6 @@ poller_modules:
     ntp: false
     ospf: false
     stp: false
-    ucd-diskio: false
     ucd-mib: false
     ipmi: false
     bgp-peers: false
