+++ resources/definitions/os_detection/stulz.yaml
@@ -14,7 +14,6 @@ discovery_modules:
     mpls: false
     storage: false
     stp: false
-    ucd-diskio: false
     vlans: false
     wireless: false
 discovery:
